Ensure table view cell has higher zindex then footer.
This is a bit hacky and should probably be done a bit
differently. For now, it shows off the overall design.

Footers in Table Views will potentially staticly affix
it self during scrolling (hence putting TableViewCells
above via zIndex).

The improvement would be to not use Footers on the
Table View and instead modify the last cell in each
section to build it's own rounding.